Auto populate all of the sensor classes
The generation is based upon contents of the sensors directory. To
create a new sensor you just define a new .py file in sensors directory
that includes a single class with a name that matches the file name.
Then you simply specify the type of the sensor in the config as that
name.

CR: me, myself, I